protein_tracker2/requirements.txt
flask==3.0.3
anthropic==0.34.2
python-dotenv==1.0.1
Pillow==10.4.0
google-genai>=1.0.0
Flask-Login
